chore: renaming config package
