+++ backend/src/main/java/ca/bc/gov/restapi/results/common/configuration/SilvaHikariConfig.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.restapi.results.common.config;
+package ca.bc.gov.restapi.results.common.configuration;
 
 import lombok.Getter;
 import lombok.Setter;
