+++ backend/src/main/java/ca/bc/gov/restapi/results/oracle/configuration/OracleJpaConfig.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.restapi.results.oracle.config;
+package ca.bc.gov.restapi.results.oracle.configuration;
 
 import ca.bc.gov.restapi.results.oracle.entity.CutBlockOpenAdminEntity;
 import ca.bc.gov.restapi.results.oracle.entity.OpeningEntity;
